refactor: modularizar app.py em pacote app/ com Blueprints
- Separar 1.734 linhas em 14 arquivos com responsabilidades definidas
- App factory pattern (create_app) para facilitar testes
- Flask Blueprints para cada domínio: auth, lancamentos, orcamentos, metas, relatorios, contas, social, pluggy, ia, importacao
- Extrair config, database e utils em módulos próprios
- Atualizar testes para nova estrutura
- 50 testes passando sem alteração de comportamento
